rewrite core
